chore: remove changes from sdk and put it in demo-app
