+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/notifications/internal/telecom/TelecomCallController.kt
@@ -123,7 +123,7 @@ class TelecomCallController(val context: Context) {
                     }
                 }
             } else {
-                //TODO Rahul Seems wrong should be telecomServiceLauncher.addOnGoingCall()
+                // TODO Rahul Seems wrong should be telecomServiceLauncher.addOnGoingCall()
                 telecomServiceLauncher.addOutgoingCallToTelecom(
                     context,
                     callId = StreamCallId(call.type, call.id),
