+++ stream-video-android-ui-compose/src/main/kotlin/io/getstream/video/android/compose/ui/components/call/activecall/AudioCallContent.kt
@@ -29,16 +29,10 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.BluetoothAudio
-import androidx.compose.material.icons.filled.Headphones
-import androidx.compose.material.icons.filled.HeadsetMic
-import androidx.compose.material.icons.filled.SpeakerPhone
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
@@ -53,16 +47,14 @@ import io.getstream.video.android.compose.ui.components.call.activecall.internal
 import io.getstream.video.android.compose.ui.components.call.controls.actions.DefaultOnCallActionHandler
 import io.getstream.video.android.compose.ui.components.call.controls.actions.LeaveCallAction
 import io.getstream.video.android.compose.ui.components.call.controls.actions.ToggleMicrophoneAction
-import io.getstream.video.android.compose.ui.components.call.ringing.outgoingcall.MicSelectorDropDown
 import io.getstream.video.android.compose.ui.components.call.ringing.outgoingcall.OutgoingCallContent
-import io.getstream.video.android.compose.ui.components.call.ringing.outgoingcall.OutgoingCallControlsV2
+import io.getstream.video.android.compose.ui.components.call.ringing.outgoingcall.OutgoingCallControls
 import io.getstream.video.android.compose.ui.components.participants.ParticipantAvatars
 import io.getstream.video.android.compose.ui.components.participants.internal.ParticipantInformation
 import io.getstream.video.android.core.Call
 import io.getstream.video.android.core.MemberState
 import io.getstream.video.android.core.ParticipantState
 import io.getstream.video.android.core.StreamVideo
-import io.getstream.video.android.core.audio.StreamAudioDevice
 import io.getstream.video.android.core.call.state.CallAction
 import io.getstream.video.android.core.model.CallStatus
 import io.getstream.video.android.core.telecom.TelecomPermissions
@@ -137,36 +129,15 @@ public fun AudioCallContent(
         },
         onBackPressed = onBackPressed,
         controlsContent = controlsContent ?: {
-            val availableDevices by call.microphone.devices.collectAsStateWithLifecycle()
-            val selectedMicroPhoneDevice by call.microphone.selectedDevice.collectAsStateWithLifecycle()
-            val audioDeviceUiStateList: List<AudioDeviceUiState> = availableDevices.map {
-                val icon = when (it) {
-                    is StreamAudioDevice.BluetoothHeadset -> Icons.Default.BluetoothAudio
-                    is StreamAudioDevice.Earpiece -> Icons.Default.Headphones
-                    is StreamAudioDevice.Speakerphone -> Icons.Default.SpeakerPhone
-                    is StreamAudioDevice.WiredHeadset -> Icons.Default.HeadsetMic
-                }
-                AudioDeviceUiState(
-                    it,
-                    it.name,
-                    icon,
-                    it.name == selectedMicroPhoneDevice?.name,
-                )
-            }
-
-            if (selectedMicroPhoneDevice != null) {
-                OutgoingCallControlsV2(
-                    modifier = Modifier
-                        .align(Alignment.BottomCenter)
-                        .padding(VideoTheme.dimens.spacingM),
-                    isVideoCall = false,
-                    isCameraEnabled = false,
-                    isMicrophoneEnabled = isMicrophoneEnabled,
-                    audioDeviceUiStateList = audioDeviceUiStateList,
-                    call = call,
-                    onCallAction = onCallAction,
-                )
-            }
+            OutgoingCallControls(
+                modifier = Modifier
+                    .align(Alignment.BottomCenter)
+                    .padding(VideoTheme.dimens.spacingM),
+                isVideoCall = false,
+                isCameraEnabled = false,
+                isMicrophoneEnabled = isMicrophoneEnabled,
+                onCallAction = onCallAction,
+            )
         },
     )
 }
@@ -266,30 +237,12 @@ public fun AudioOnlyCallContent(
                     duration = durationText,
                 )
         }
-        val selectedMicroPhoneDevice by call.microphone.selectedDevice.collectAsStateWithLifecycle()
-        val availableDevices by call.microphone.devices.collectAsStateWithLifecycle()
-        val audioDeviceUiStateList: List<AudioDeviceUiState> = availableDevices.map {
-            val icon = when (it) {
-                is StreamAudioDevice.BluetoothHeadset -> Icons.Default.BluetoothAudio
-                is StreamAudioDevice.Earpiece -> Icons.Default.Headphones
-                is StreamAudioDevice.Speakerphone -> Icons.Default.SpeakerPhone
-                is StreamAudioDevice.WiredHeadset -> Icons.Default.HeadsetMic
-            }
-            AudioDeviceUiState(
-                it,
-                it.name,
-                icon,
-                it.name == selectedMicroPhoneDevice?.name,
-            )
-        }
 
-        controlsContent?.invoke(this) ?: AudioOnlyCallControlsV2(
+        controlsContent?.invoke(this) ?: AudioOnlyCallControls(
             modifier = Modifier
                 .align(Alignment.BottomCenter)
                 .padding(bottom = VideoTheme.dimens.genericXxl),
             isMicrophoneEnabled = isMicrophoneEnabled,
-            audioDeviceUiStateList = audioDeviceUiStateList,
-            call = call,
             onCallAction = onCallAction,
         )
     }
@@ -328,7 +281,6 @@ public fun AudioOnlyCallDetails(
  * @param isMicrophoneEnabled Weather or not the microphone icon will show the mic as enabled or not.
  * @param onCallAction Handler used when the user triggers a [CallAction].
  */
-@Deprecated("Use the new AudioOnlyCallControlsV2 instead.", level = DeprecationLevel.ERROR)
 @Composable
 public fun AudioOnlyCallControls(
     modifier: Modifier,
@@ -347,39 +299,6 @@ public fun AudioOnlyCallControls(
             offStyle = VideoTheme.styles.buttonStyles.secondaryIconButtonStyle().fillCircle(1.5f),
             onStyle = VideoTheme.styles.buttonStyles.tertiaryIconButtonStyle().fillCircle(1.5f),
         )
-
-        LeaveCallAction(
-            modifier = Modifier.testTag("Stream_HangUpButton"),
-            onCallAction = onCallAction,
-            style = VideoTheme.styles.buttonStyles.primaryIconButtonStyle().fillCircle(1.5f),
-        )
-    }
-}
-
-@Composable
-public fun AudioOnlyCallControlsV2(
-    modifier: Modifier,
-    isMicrophoneEnabled: Boolean,
-    call: Call,
-    audioDeviceUiStateList: List<AudioDeviceUiState>,
-    onCallAction: (CallAction) -> Unit,
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceEvenly,
-    ) {
-        ToggleMicrophoneAction(
-            modifier = Modifier.testTag("Stream_MicrophoneToggle_Enabled_$isMicrophoneEnabled"),
-            isMicrophoneEnabled = isMicrophoneEnabled,
-            onCallAction = onCallAction,
-            offStyle = VideoTheme.styles.buttonStyles.secondaryIconButtonStyle().fillCircle(1.5f),
-            onStyle = VideoTheme.styles.buttonStyles.tertiaryIconButtonStyle().fillCircle(1.5f),
-        )
-
-        if (audioDeviceUiStateList.isNotEmpty()) {
-            MicSelectorDropDown(call, audioDeviceUiStateList)
-        }
 
         LeaveCallAction(
             modifier = Modifier.testTag("Stream_HangUpButton"),
@@ -415,10 +334,3 @@ private fun AudioOnlyCallContentPreview() {
         )
     }
 }
-
-public data class AudioDeviceUiState(
-    val streamAudioDevice: StreamAudioDevice,
-    val text: String,
-    val icon: ImageVector, // Assuming it's a drawable resource ID
-    val highlight: Boolean,
-)
